test: obj_ulog_size fix counter type
Ref: pmem/issues#3866

+++ src/test/obj_ulog_size/obj_ulog_size.c
@@ -79,7 +79,7 @@
 static void
 free_pool(PMEMoid *oids, size_t noids)
 {
-	for (int i = 0; i < noids; i++) {
+	for (size_t i = 0; i < noids; i++) {
 		pmemobj_free(&oids[i]);
 		UT_ASSERT(OID_IS_NULL(oids[i]));
 	}
